Add a11y-support to "Show solution"-dialog
Clicking the "Show solution"-button will now identify the next slide as
a dialog.

Punctuation is added to every question and answer that misses it, to
ensure that Read Speakers doesn't combine different questions and
answers to a single unit.

Adds focus to the "Retry"-button after closing the Solution-dialog.

Resolves: #HFP-95